Fixes seeder and tests
